Align the module with new Origami standards
Some of the tests in here are a little convoluted, because we have to
mock the global `Date` function. Sinon doesn't provide mocking of dates
in the same way as Jasmine so I've had to work around that.

I'm going to rewrite and simplify a lot of these in a later PR.